mv hyper_api_images_test.go to passed and future
